+++ src/main/java/com/example/springboot_education/repositories/AssignmentJpaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AssignmentJpaRepository extends JpaRepository<Assignment, Long> {
+public interface AssignmentJpaRepository extends JpaRepository<Assignment, Integer> {
 }
